+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+### 2.0.0 (unreleased)
+
+ * ability to open a rucksack bundle which is a 0 byte empty file.
+ * added a setting which allows preventing rotating images 90 degrees.
+ * ability to glob for images in a texture.
+ * (API breaking) rename all `page` functions to `texture`.
+ * (API breaking) remove `RuckSackImage` property `name` in favor of `key`,
+   `key_size`, and `path`.
+
 ### 1.0.1
 
  * Update docs and add changelog